+++ packages/core/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 1.7.1 (2021-05-03)
 - Use `web3_clientVersion` to better detect development networks.
 
 ## 1.7.0 (2021-04-29)